Update font file path and base directory handling
Changed file path in 'embed-fonts.config.json' to refer to local node module instead of an external URL. Also, updated the 'index.js' to improve the handling of base directory. The modifications were made in order to streamline the flow of fetching CSS stylesheets and resolving font file paths. The base directory is now reset and updated for each file processed, and the font file path will only be resolved to an absolute path if it doesn't start with 'http'. This allows for more flexible handling of relative paths and URLs.